Detach RecentlyAddedDocumentList.js from DashboardPage.js
